requirements.txt
langchain==0.3.9
transformers==4.46.3
torch==2.2.0+cpu
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.10.0
numpy==1.24.2
tqdm==4.66.2
bitsandbytes==0.41.1
accelerate==0.27.2
